updated script to be able to run regression
