+++ lib/cv_dv_utils/python/sim_cmd/run_test.py
@@ -21,6 +21,7 @@
 import argparse
 import os
 import yaml
+import get_cmd as sim
 
 parser = argparse.ArgumentParser(description='compile options')
 parser.add_argument('--yaml'     ,dest='yaml_file', type=str, help='Top YAML with compile and simulation options')
@@ -35,93 +36,17 @@ parser.add_argument('--outdir'   , dest='outdir', type=str, help='output dirctor
 
 args = parser.parse_args()
 
-def get_cmd_opt(yaml_file):
-  with open(yaml_file, 'r') as yaml_top:
-     sim_yaml = yaml.safe_load(yaml_top)
-  
-  for entry in sim_yaml: 
-      if entry['tool'] == "questa":
-          cmd       = "vsim"
-          tool      = "questa"
-      elif entry['tool'] == "vcs":
-          cmd       = "./vcs"
-          tool      = "vcs"
-      if 'sim' in entry:
-        sim     = entry['sim']
-      else: 
-        sim = ""
-
-      if 'sim_option' in sim:
-        sim_opt = sim['sim_option']
-      else:
-        sim_opt = ""
-
-  sim_cmd_opt = ""
-  if tool == "questa":
-    sim_cmd_opt = "{} {}".format(cmd, sim_opt)
-    return sim_cmd_opt
-
-def run_test(test_name, seed, debug, batch, dump, coverage, stdout, outdir, vsim_opt):
-
-   if outdir == None: 
-      outdir = "output"
-
-   
-   if seed == None: 
-      seed = 1
-   
-   if debug == None: 
-      debug = "UVM_LOW"
-  
-   if batch == None: 
-      batch = 1 
-   
-   if dump == None: 
-      dump = 1 
-   
-   if batch == 1:
-     batchstr = "-c"
-   else:
-     batchstr = "-visualizer"
-   
-   if dump == 1:
-     wlfstr  = "-wlf {}/{}_{}.wlf".format(outdir, test_name, seed)
-     dumpstr = ""
-   else:
-     dumpt = ""
-     wlfstr  = ""
-
-   if stdout == None:
-       stdout = 1
-
-   if stdout == 0:
-       stdoutstr = ">"
-   else:
-       stdoutstr = "-l"
-  
-   if os.path.isdir("{}".format(outdir)) == False:
-     os.system("mkdir {}".format(outdir))
-   
-   if coverage == 1:
-     ucdbfile    = "\"coverage save -onexit {}/{}_{}.ucdb\"".format(outdir, test_name, seed)
-     coveragestr = "-do {}".format(ucdbfile)
-     coverageopt = "-coverage -assertdebug"
-   else:
-     coveragestr = ""
-     coverageopt = ""
-
-   os.system("{} {} {} {} -sv_seed {} +UVM_VERBOSITY={} +UVM_TESTNAME={} {} {}/{}_{}.log {}".format(vsim_opt,  batchstr, dumpstr, coverageopt, seed, debug, test_name, stdoutstr, outdir, test_name, seed, wlfstr))
 
 option_ok = 1
 if args.yaml_file == None:
    print("Please specify the YAML")
    option_ok = 0
 else:
-   cmd_opt = get_cmd_opt(args.yaml_file)
+   cmd_opt = sim.get_cmd_opt(args.yaml_file)
 if args.test_name == None: 
    print("Please specify the testname")
    option_ok = 0
 
 if option_ok == 1:
-  run_test(args.test_name, args.seed, args.debug, args.batch, args.dump, args.coverage, args.stdout, args.outdir, cmd_opt )
+  sim.run_test(args.test_name, args.seed, args.debug, args.batch, args.dump, args.coverage, args.stdout, args.outdir, cmd_opt )
